move custome exception back to errors.py
